+++ packages/utils/src/types.ts
@@ -1,1 +1,3 @@
 export type BrandedString<T> = string & { __brand: T }
+
+export type Events<T> = { [K in keyof T]: (payload: any) => void }
